requirements.txt
httpx
python-jose[cryptography]

# Web Framework
fastapi
uvicorn[standard]

# python adapter
mangum

# Database Toolkit (ORM)
sqlalchemy>=2.0.0

# Async PostgreSQL Driver
asyncpg

# Environment Variable Management
python-dotenv
pydantic-settings